feat(ui,api): ✨ 统一 Long 类型 ID 并迁移服务实例管理页面
将所有服务实例、角色组、角色实例等核心实体的 ID 由 Integer 升级为 Long 以支撑大数据量级。
完成 datasophon-ui InstancesTab 组件功能整体迁移，支持分页、筛选、批量操作及自动伸缩控制。

涉及文件：
- Java Service/Controller：serviceInstanceId、roleGroupId 等字段类型调整
- DAO Mapper：动态 SQL 查询条件类型同步
- UI：instances-tab.tsx 全套重构 + api-config 端点补充

BREAKING CHANGE：API 参数及 DTO 中的整型字段全部改为 Long。

+++ datasophon-api/src/main/java/com/datasophon/api/controller/v1/cluster/ClusterServiceInstanceConfigController.java
@@ -84,7 +84,7 @@ public class ClusterServiceInstanceConfigController {
     public Result<Object> info(
             @RequestParam("serviceInstanceId") Long serviceInstanceId,
             @RequestParam("version") Integer version, 
-            @RequestParam("roleGroupId") Integer roleGroupId, 
+            @RequestParam("roleGroupId") Long roleGroupId, 
             @RequestParam("page") Integer page,
             @RequestParam("pageSize") Integer pageSize) {
