+++ datasophon-common/src/main/java/com/datasophon/common/command/GenerateStarRocksHAMessage.java
@@ -24,7 +24,7 @@ import lombok.Data;
 @Data
 public class GenerateStarRocksHAMessage implements Serializable {
 
-    private Integer serviceInstanceId;
+    private Long serviceInstanceId;
 
     private Long clusterId;
 }
