+++ datasophon-service/src/main/java/com/datasophon/api/strategy/RMHandlerStrategy.java
@@ -138,7 +138,7 @@ public class RMHandlerStrategy extends ServiceHandlerAbstract implements Service
 
         @Override
         protected ConnectionInfo.ConnectionInfoBuilder getServiceSpecificConnectionInfo(
-                        Long clusterId, Integer serviceInstanceId, Map<String, String> configMap) {
+                        Long clusterId, Long serviceInstanceId, Map<String, String> configMap) {
                 try {
                         logger.info("开始获取ResourceManager服务连接信息，集群ID: {}, 服务实例ID: {}", clusterId, serviceInstanceId);
